skip install during quilt setup run
When quilt runs, it sets env variable QUILT_COMMAND to "setup".
This happens during rpm package unpack, for example. We can detect
this step and not run the script

This allows the proxy to be run directly without the helper bash
script, so you can run with preferred node version.

Closes: https://github.com/openSUSE/local-npm-registry/issues/4